requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0


